requirements.txt
annotated-types==0.7.0
anyio==4.9.0
Authlib==1.6.1
bcrypt==4.3.0
certifi==2025.7.14
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.116.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
passlib==1.7.4
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyMySQL==1.1.1
python-dotenv==1.1.1
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
